Various fixes and README addition.
Now uses hashes on the js and css assets so that they always show the latest changes when you deploy them. PFP has moved to src/public/images so it doesn't get hashed. Added LICENSE.txt and README.md. Removed cypress as a dependency as we are not testing anything here.